+++ nativelink-store/src/redis_store.rs
@@ -34,8 +34,8 @@ use uuid::Uuid;
 use crate::cas_utils::is_zero_digest;
 
 // TODO(caass): These (and other settings) should be made configurable via nativelink-config.
-pub const READ_CHUNK_SIZE: usize = 64 * 1024;
-const CONNECTION_POOL_SIZE: usize = 3;
+pub const READ_CHUNK_SIZE: u64 = 64 * 1024;
+const CONNECTION_POOL_SIZE: u64 = 3;
 
 /// A [`StoreDriver`] implementation that uses Redis as a backing store.
 #[derive(MetricsComponent)]
@@ -119,7 +119,7 @@ impl RedisStore {
                 broadcast_channel_capacity: 4096,
                 ..Default::default()
             })
-            .build_pool(CONNECTION_POOL_SIZE)
+            .build_pool(CONNECTION_POOL_SIZE as usize)
             .err_tip(|| "while creating redis connection pool")?;
 
         let subscriber_client = builder
@@ -178,7 +178,7 @@ impl StoreDriver for RedisStore {
     async fn has_with_results(
         self: Pin<&Self>,
         keys: &[StoreKey<'_>],
-        results: &mut [Option<usize>],
+        results: &mut [Option<u64>],
     ) -> Result<(), Error> {
         // TODO(caass): Optimize for the case where `keys.len() == 1`
         let pipeline = self.client_pool.next().pipeline();
@@ -334,8 +334,8 @@ impl StoreDriver for RedisStore {
         self: Pin<&Self>,
         key: StoreKey<'_>,
         writer: &mut DropCloserWriteHalf,
-        offset: usize,
-        length: Option<usize>,
+        offset: u64,
+        length: Option<u64>,
     ) -> Result<(), Error> {
         // To follow RBE spec we need to consider any digest's with
         // zero size to be existing.
@@ -355,7 +355,7 @@ impl StoreDriver for RedisStore {
         // We want to read the data at the key from `offset` to `offset + length`.
         let data_start = offset;
         let data_end = data_start
-            .saturating_add(length.unwrap_or(isize::MAX as usize))
+            .saturating_add(length.unwrap_or(isize::MAX as u64))
             .saturating_sub(1);
 
         // And we don't ever want to read more than `READ_CHUNK_SIZE` bytes at a time, so we'll need to iterate.
@@ -364,11 +364,11 @@ impl StoreDriver for RedisStore {
 
         loop {
             let chunk: Bytes = client
-                .getrange(encoded_key, chunk_start, chunk_end)
+                .getrange(encoded_key, chunk_start as usize, chunk_end as usize)
                 .await
                 .err_tip(|| "In RedisStore::get_part::getrange")?;
 
-            let didnt_receive_full_chunk = chunk.len() < READ_CHUNK_SIZE;
+            let didnt_receive_full_chunk = chunk.len() < READ_CHUNK_SIZE as usize;
             let reached_end_of_data = chunk_end == data_end;
 
             if didnt_receive_full_chunk || reached_end_of_data {
